+++ sdk/src/DocuSign.eSign/Api/BulkEnvelopesApi.cs
@@ -131,7 +131,7 @@ namespace DocuSign.eSign.Api
         /// Updates the bulk recipients in a draft envelope using a file upload. The Content-Type supported for uploading a bulk recipient file is CSV (text/csv).  The REST API does not support modifying individual rows or values in the bulk recipients file. It only allows the entire file to be added or replaced with a new file.
         /// </remarks>
         /// <exception cref="DocuSign.eSign.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="accountId">The external account number (int) or account ID Guid.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
+        /// <param name="accountId">The external account number (int) or account id GUID.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
         /// <param name="bulkRecipientsRequest"></param>
         
         /// <returns></returns>
@@ -144,7 +144,7 @@ namespace DocuSign.eSign.Api
         /// Updates the bulk recipients in a draft envelope using a file upload. The Content-Type supported for uploading a bulk recipient file is CSV (text/csv).  The REST API does not support modifying individual rows or values in the bulk recipients file. It only allows the entire file to be added or replaced with a new file.
         /// </remarks>
         /// <exception cref="DocuSign.eSign.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="accountId">The external account number (int) or account ID Guid.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
+        /// <param name="accountId">The external account number (int) or account id GUID.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
         /// <param name="bulkRecipientsRequest"></param>
         
         /// <returns>ApiResponse of </returns>
@@ -258,7 +258,7 @@ namespace DocuSign.eSign.Api
         /// Updates the bulk recipients in a draft envelope using a file upload. The Content-Type supported for uploading a bulk recipient file is CSV (text/csv).  The REST API does not support modifying individual rows or values in the bulk recipients file. It only allows the entire file to be added or replaced with a new file.
         /// </remarks>
         /// <exception cref="DocuSign.eSign.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="accountId">The external account number (int) or account ID Guid.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
+        /// <param name="accountId">The external account number (int) or account id GUID.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
         /// <param name="bulkRecipientsRequest"></param>
         
         /// <returns>Task of BulkRecipientsSummaryResponse</returns>
@@ -271,7 +271,7 @@ namespace DocuSign.eSign.Api
         /// Updates the bulk recipients in a draft envelope using a file upload. The Content-Type supported for uploading a bulk recipient file is CSV (text/csv).  The REST API does not support modifying individual rows or values in the bulk recipients file. It only allows the entire file to be added or replaced with a new file.
         /// </remarks>
         /// <exception cref="DocuSign.eSign.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="accountId">The external account number (int) or account ID Guid.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
+        /// <param name="accountId">The external account number (int) or account id GUID.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
         /// <param name="bulkRecipientsRequest"></param>
         
         /// <returns>Task of ApiResponse (BulkRecipientsSummaryResponse)</returns>
@@ -1150,7 +1150,7 @@ namespace DocuSign.eSign.Api
         /// Adds or replaces envelope bulk recipients. Updates the bulk recipients in a draft envelope using a file upload. The Content-Type supported for uploading a bulk recipient file is CSV (text/csv).  The REST API does not support modifying individual rows or values in the bulk recipients file. It only allows the entire file to be added or replaced with a new file.
         /// </summary>
         /// <exception cref="DocuSign.eSign.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="accountId">The external account number (int) or account ID Guid.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
+        /// <param name="accountId">The external account number (int) or account id GUID.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
         /// <param name="bulkRecipientsRequest"></param>
         
         /// <returns>BulkRecipientsSummaryResponse</returns>
@@ -1164,7 +1164,7 @@ namespace DocuSign.eSign.Api
         /// Adds or replaces envelope bulk recipients. Updates the bulk recipients in a draft envelope using a file upload. The Content-Type supported for uploading a bulk recipient file is CSV (text/csv).  The REST API does not support modifying individual rows or values in the bulk recipients file. It only allows the entire file to be added or replaced with a new file.
         /// </summary>
         /// <exception cref="DocuSign.eSign.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="accountId">The external account number (int) or account ID Guid.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
+        /// <param name="accountId">The external account number (int) or account id GUID.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
         /// <param name="bulkRecipientsRequest"></param>
         
         /// <returns>ApiResponse of BulkRecipientsSummaryResponse</returns>
@@ -1253,7 +1253,7 @@ namespace DocuSign.eSign.Api
         /// Adds or replaces envelope bulk recipients. Updates the bulk recipients in a draft envelope using a file upload. The Content-Type supported for uploading a bulk recipient file is CSV (text/csv).  The REST API does not support modifying individual rows or values in the bulk recipients file. It only allows the entire file to be added or replaced with a new file.
         /// </summary>
         /// <exception cref="DocuSign.eSign.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="accountId">The external account number (int) or account ID Guid.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
+        /// <param name="accountId">The external account number (int) or account id GUID.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
         /// <param name="bulkRecipientsRequest"></param>
         
         /// <returns>Task of BulkRecipientsSummaryResponse</returns>
@@ -1268,7 +1268,7 @@ namespace DocuSign.eSign.Api
         /// Adds or replaces envelope bulk recipients. Updates the bulk recipients in a draft envelope using a file upload. The Content-Type supported for uploading a bulk recipient file is CSV (text/csv).  The REST API does not support modifying individual rows or values in the bulk recipients file. It only allows the entire file to be added or replaced with a new file.
         /// </summary>
         /// <exception cref="DocuSign.eSign.Client.ApiException">Thrown when fails to make API call</exception>
-        /// <param name="accountId">The external account number (int) or account ID Guid.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
+        /// <param name="accountId">The external account number (int) or account id GUID.</param>/// <param name="envelopeId">The envelope&#39;s GUID. Eg 93be49ab-afa0-4adf-933c-f752070d71ec </param>/// <param name="recipientId">The &#x60;recipientId&#x60; used when the envelope or template was created.</param>
         /// <param name="bulkRecipientsRequest"></param>
         
         /// <returns>Task of ApiResponse (BulkRecipientsSummaryResponse)</returns>