+++ sdk/src/DocuSign.eSign/Model/EnvelopeDocument.cs
@@ -51,9 +51,10 @@ namespace DocuSign.eSign.Model
         /// <param name="Order">.</param>
         /// <param name="Pages">.</param>
         /// <param name="SignerMustAcknowledge">.</param>
+        /// <param name="SizeBytes">.</param>
         /// <param name="Type">.</param>
         /// <param name="Uri">.</param>
-        public EnvelopeDocument(string AttachmentTabId = default(string), string AuthoritativeCopy = default(string), List<SignatureType> AvailableDocumentTypes = default(List<SignatureType>), string ContainsPdfFormFields = default(string), string Display = default(string), List<NameValue> DocumentFields = default(List<NameValue>), string DocumentGroup = default(string), string DocumentId = default(string), ErrorDetails ErrorDetails = default(ErrorDetails), string IncludeInDownload = default(string), string Name = default(string), string Order = default(string), string Pages = default(string), string SignerMustAcknowledge = default(string), string Type = default(string), string Uri = default(string))
+        public EnvelopeDocument(string AttachmentTabId = default(string), string AuthoritativeCopy = default(string), List<SignatureType> AvailableDocumentTypes = default(List<SignatureType>), string ContainsPdfFormFields = default(string), string Display = default(string), List<NameValue> DocumentFields = default(List<NameValue>), string DocumentGroup = default(string), string DocumentId = default(string), ErrorDetails ErrorDetails = default(ErrorDetails), string IncludeInDownload = default(string), string Name = default(string), string Order = default(string), string Pages = default(string), string SignerMustAcknowledge = default(string), string SizeBytes = default(string), string Type = default(string), string Uri = default(string))
         {
             this.AttachmentTabId = AttachmentTabId;
             this.AuthoritativeCopy = AuthoritativeCopy;
@@ -69,6 +70,7 @@ namespace DocuSign.eSign.Model
             this.Order = Order;
             this.Pages = Pages;
             this.SignerMustAcknowledge = SignerMustAcknowledge;
+            this.SizeBytes = SizeBytes;
             this.Type = Type;
             this.Uri = Uri;
         }
@@ -160,6 +162,12 @@ namespace DocuSign.eSign.Model
         /// 
         /// </summary>
         /// <value></value>
+        [DataMember(Name="sizeBytes", EmitDefaultValue=false)]
+        public string SizeBytes { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <value></value>
         [DataMember(Name="type", EmitDefaultValue=false)]
         public string Type { get; set; }
         /// <summary>
@@ -190,6 +198,7 @@ namespace DocuSign.eSign.Model
             sb.Append("  Order: ").Append(Order).Append("\n");
             sb.Append("  Pages: ").Append(Pages).Append("\n");
             sb.Append("  SignerMustAcknowledge: ").Append(SignerMustAcknowledge).Append("\n");
+            sb.Append("  SizeBytes: ").Append(SizeBytes).Append("\n");
             sb.Append("  Type: ").Append(Type).Append("\n");
             sb.Append("  Uri: ").Append(Uri).Append("\n");
             sb.Append("}\n");
@@ -299,6 +308,11 @@ namespace DocuSign.eSign.Model
                     this.SignerMustAcknowledge.Equals(other.SignerMustAcknowledge)
                 ) && 
                 (
+                    this.SizeBytes == other.SizeBytes ||
+                    this.SizeBytes != null &&
+                    this.SizeBytes.Equals(other.SizeBytes)
+                ) && 
+                (
                     this.Type == other.Type ||
                     this.Type != null &&
                     this.Type.Equals(other.Type)
@@ -349,6 +363,8 @@ namespace DocuSign.eSign.Model
                     hash = hash * 59 + this.Pages.GetHashCode();
                 if (this.SignerMustAcknowledge != null)
                     hash = hash * 59 + this.SignerMustAcknowledge.GetHashCode();
+                if (this.SizeBytes != null)
+                    hash = hash * 59 + this.SizeBytes.GetHashCode();
                 if (this.Type != null)
                     hash = hash * 59 + this.Type.GetHashCode();
                 if (this.Uri != null)